GUI - implement very naive code syncing
There's now two buttons to run code - run and cue (placeholder names). Run executes the code in the code box immediately, cue delays execution and starts the code at the beginning of the next bar.

This is all currently hardcoded - so the BPM is 60 and the number of beats in a bar (the quantum) is 4.