Time - unify on server time
The server's time is the canonical timeline. The clients all use an NTP-like algorithm to figure out the round trip latency to then determine the delta between their timeline and the servers. This is currently measured a bunch of times at the start of the session and then once every 10 seconds afterwards and all measurements are passed through a Kalman filter to remove noise. The architecture has also been shifted slightly ready to accommodate a tagged pre-scheduler.

This means that all clients are much closer in sync and clock drift is accounted for. It also means the issue of putting a laptop to sleep and waking it up does not affect the time calculations.

There's definitely a lot of bolt-tightening and fiddling to tune up this code still required. However, it feels like a huge leap in the right direction.