Content - use lua to structure and represent content
we now read in the content as a lua file rather than having it be directly embedded in the elixir. This means we can create nice utility functions such as `markdown` and `editor.

This is a step in the direction of making it easy to author and work with multiple sets of content.