GUI - make the cue button idempotent for a given bar
Multiple clicks of the cue button do not generate multiple playing cue-runs playing concurrently in the same bar. Hitting cue will only schedule at most one cue-run for the next bar.

This is slightly more sophisticated than hitting stop prior to cue as that would stop any existing cues that are currently playing. The idempotency is therefore associated with the next bar which if currently has a cue scheduled is de-scheduled and replaced.

Runs (via the Run button) are unaffected and are not de-scheduled by hitting cue and may still be layered up. This is achieved by differentiating between runs created by the run button and runs created by the cue button.