Lang - automatically pop all FX at the end of a script
This is currently very hacky as it waits 10 beats before popping all FX which may cut off the audio if it persists 10 beats longer than the last thing that was exectued in the script (i.e. a long running sample).

We need a better way of determining when all audio generated by a script has completed and to use that to auto pop_all_fx